+++ tests/temp_tests/test_settings/conftest.py
@@ -26,17 +26,13 @@
 
 import pytest
 
-from smartsim._core.utils.launcher import (
-    ExecutableProtocol,
-    LauncherProtocol,
-    create_job_id,
-)
+from smartsim._core.utils.launcher import LauncherProtocol, create_job_id
 from smartsim.settings.arguments import launchArguments as launch
 
 
 @pytest.fixture
 def mock_echo_executable():
-    class _MockExe(ExecutableProtocol):
+    class _MockExe:
         def as_program_arguments(self):
             return ("echo", "hello", "world")
