+++ tests/temp_tests/test_settings/test_localLauncher.py
@@ -152,12 +152,12 @@ def test_format_env_vars():
     ]
 
 
-def test_formatting_returns_original_exe(mock_echo_executable, test_dir):
+def test_formatting_returns_original_exe(test_dir):
     out = os.path.join(test_dir, "out.txt")
     err = os.path.join(test_dir, "err.txt")
     open(out, "w"), open(err, "w")
     shell_launch_cmd = _as_local_command(
-        LocalLaunchArguments({}), mock_echo_executable, test_dir, {}, out, err
+        LocalLaunchArguments({}), ("echo", "hello", "world"), test_dir, {}, out, err
     )
     assert isinstance(shell_launch_cmd, ShellLauncherCommand)
     assert shell_launch_cmd.command_tuple == ("echo", "hello", "world")
