+++ tests/test_shell_launcher.py
@@ -46,7 +46,7 @@ class EchoHelloWorldEntity(entity.SmartSimEntity):
     """A simple smartsim entity that meets the `ExecutableProtocol` protocol"""
 
     def __init__(self):
-        super().__init__("test-entity", "echo", None, None)
+        super().__init__("test-entity")
 
     def __eq__(self, other):
         if type(self) is not type(other):
